Swap tests around so each file is runnable on its own
